+++ shiny/experimental/ui/_deprecated.py
@@ -60,11 +60,10 @@ from ...ui._sidebar import panel_main as main_panel_main
 from ...ui._sidebar import panel_sidebar as main_panel_sidebar
 from ...ui._sidebar import sidebar as main_sidebar
 from ...ui._sidebar import toggle_sidebar as main_toggle_sidebar
-from ...ui.css_unit._css_unit import CssUnit as MainCssUnit
-from ...ui.css_unit._css_unit import as_css_padding as main_as_css_padding
-from ...ui.css_unit._css_unit import as_css_unit as main_as_css_unit
-from ...ui.css_unit._css_unit import as_width_unit as main_as_width_unit
-from ...ui.fill import as_fill_carrier as main_as_fill_carrier
+from ...ui.css._css_unit import CssUnit as MainCssUnit
+from ...ui.css._css_unit import as_css_padding as main_as_css_padding
+from ...ui.css._css_unit import as_css_unit as main_as_css_unit
+from ...ui.css._css_unit import as_width_unit as main_as_width_unit
 from ...ui.fill import as_fill_item as main_as_fill_item
 from ...ui.fill import as_fillable_container as main_as_fillable_container
 from ...ui.fill import is_fill_item as main_is_fill_item