+++ shiny/ui/_navs.py
@@ -33,7 +33,7 @@ from ._card import CardItem, card, card_body, card_footer, card_header
 from ._html_deps_shinyverse import nav_spacer_dependency
 from ._sidebar import Sidebar, layout_sidebar
 from ._tag import tag_add_style
-from .css_unit import CssUnit, as_css_padding, as_css_unit
+from .css import CssUnit, as_css_padding, as_css_unit
 from .fill import as_fill_item, as_fillable_container